Use Salesforce to read data (#3)
Before, the member data had to be loaded via a CSV that we manually
exported. Now, we dynamically read in the data through the Salesforce
REST API and the helper library
https://github.com/simple-salesforce/simple-salesforce.

A follow-up will save the results to Salesforce.

For now, I turned off Mailchimp data. Our CSV is out-of-date and all
those entries were already processed and saved in Salesforce a while
ago.

Note that Salesforce stores unset values as `None`, rather than `""`, so
we change `SalesforceEntry` to better model unset values.