Eagerly write updates to Salesforce (#12)
I was about 800 records through updates when the program crashed because
of an unrecognized state code. This is the second time a crash has
happened. The issue is that none of the prior changes are saved in this
case, so we have to rerun the entire program from the start.

Now, the program eagerly writes updates when they are computed. Even if
we crash on record 1000, the first 999 will have been written.